plaecholders for CGI functions, temporarily passing request to request handler
