+++ inc/CgiHandler.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "StandardLibraries.hpp"
+
+class CgiHandler {
+public:
+	int pid;
+	void	launchCgi();
+	void	checkCgi();
+};
