+++ src/CgiHandler.cpp
@@ -1,0 +1,6 @@
+#include "../inc/CgiHandler.hpp"
+
+void	CgiHandler::launchCgi()
+{}
+void	CgiHandler::checkCgi()
+{}
